Extract `reject_early_close` tests
